Add new class Component, refactor Film and FilmPopup classes
